+++ scrapy/contracts/__init__.py
@@ -3,7 +3,7 @@ import sys
 from functools import wraps
 from inspect import getmembers
 from types import CoroutineType
-from typing import AsyncGenerator, Dict
+from typing import AsyncGenerator, Dict, Optional, Type
 from unittest import TestCase
 
 from scrapy.http import Request
@@ -14,7 +14,7 @@ from scrapy.utils.spider import iterate_spider_output
 class Contract:
     """Abstract class for contracts"""
 
-    request_cls = None
+    request_cls: Optional[Type[Request]] = None
 
     def __init__(self, method, *args):
         self.testcase_pre = _create_testcase(method, f"@{self.name} pre-hook")
@@ -41,7 +41,9 @@ class Contract:
                     cb_result = cb(response, **cb_kwargs)
                     if isinstance(cb_result, (AsyncGenerator, CoroutineType)):
                         raise TypeError("Contracts don't support async callbacks")
-                    return list(iterate_spider_output(cb_result))
+                    return list(  # pylint: disable=return-in-finally
+                        iterate_spider_output(cb_result)
+                    )
 
             request.callback = wrapper
 
@@ -68,7 +70,7 @@ class Contract:
                 else:
                     results.addSuccess(self.testcase_post)
                 finally:
-                    return output
+                    return output  # pylint: disable=return-in-finally
 
             request.callback = wrapper