+++ tests/test_scheduler.py
@@ -66,7 +66,7 @@ class MockCrawler(Crawler):
 
 
 class SchedulerHandler:
-    priority_queue_cls = None
+    priority_queue_cls: Optional[str] = None
     jobdir = None
 
     def create_scheduler(self):
@@ -454,7 +454,7 @@ def _is_scheduling_fair(enqueued_slots, dequeued_slots):
 
 
 class DownloaderAwareSchedulerTestMixin:
-    priority_queue_cls = "scrapy.pqueues.DownloaderAwarePriorityQueue"
+    priority_queue_cls: Optional[str] = "scrapy.pqueues.DownloaderAwarePriorityQueue"
     reopen = False
 
     def test_logic(self):