+++ tests/test_spidermiddleware_referer.py
@@ -1,4 +1,5 @@
 import warnings
+from typing import Any, Dict, List, Optional, Tuple
 from unittest import TestCase
 from urllib.parse import urlparse
 
@@ -31,10 +32,10 @@ from scrapy.spiders import Spider
 
 
 class TestRefererMiddleware(TestCase):
-    req_meta = {}
-    resp_headers = {}
-    settings = {}
-    scenarii = [
+    req_meta: Dict[str, Any] = {}
+    resp_headers: Dict[str, str] = {}
+    settings: Dict[str, Any] = {}
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         ("http://scrapytest.org", "http://scrapytest.org/", b"http://scrapytest.org"),
     ]
 
@@ -64,7 +65,7 @@ class MixinDefault:
     with some additional filtering of s3://
     """
 
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         ("https://example.com/", "https://scrapy.org/", b"https://example.com/"),
         ("http://example.com/", "http://scrapy.org/", b"http://example.com/"),
         ("http://example.com/", "https://scrapy.org/", b"http://example.com/"),
@@ -85,7 +86,7 @@ class MixinDefault:
 
 
 class MixinNoReferrer:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         ("https://example.com/page.html", "https://example.com/", None),
         ("http://www.example.com/", "https://scrapy.org/", None),
         ("http://www.example.com/", "http://scrapy.org/", None),
@@ -95,7 +96,7 @@ class MixinNoReferrer:
 
 
 class MixinNoReferrerWhenDowngrade:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # TLS to TLS: send non-empty referrer
         (
             "https://example.com/page.html",
@@ -177,7 +178,7 @@ class MixinNoReferrerWhenDowngrade:
 
 
 class MixinSameOrigin:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # Same origin (protocol, host, port): send referrer
         (
             "https://example.com/page.html",
@@ -246,7 +247,7 @@ class MixinSameOrigin:
 
 
 class MixinOrigin:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # TLS or non-TLS to TLS or non-TLS: referrer origin is sent (yes, even for downgrades)
         (
             "https://example.com/page.html",
@@ -270,7 +271,7 @@ class MixinOrigin:
 
 
 class MixinStrictOrigin:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # TLS or non-TLS to TLS or non-TLS: referrer origin is sent but not for downgrades
         (
             "https://example.com/page.html",
@@ -298,7 +299,7 @@ class MixinStrictOrigin:
 
 
 class MixinOriginWhenCrossOrigin:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # Same origin (protocol, host, port): send referrer
         (
             "https://example.com/page.html",
@@ -405,7 +406,7 @@ class MixinOriginWhenCrossOrigin:
 
 
 class MixinStrictOriginWhenCrossOrigin:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # Same origin (protocol, host, port): send referrer
         (
             "https://example.com/page.html",
@@ -517,7 +518,7 @@ class MixinStrictOriginWhenCrossOrigin:
 
 
 class MixinUnsafeUrl:
-    scenarii = [
+    scenarii: List[Tuple[str, str, Optional[bytes]]] = [
         # TLS to TLS: send referrer
         (
             "https://example.com/sekrit.html",
@@ -920,7 +921,9 @@ class TestPolicyHeaderPrecedence004(
 
 class TestReferrerOnRedirect(TestRefererMiddleware):
     settings = {"REFERRER_POLICY": "scrapy.spidermiddlewares.referer.UnsafeUrlPolicy"}
-    scenarii = [
+    scenarii: List[
+        Tuple[str, str, Tuple[Tuple[int, str], ...], Optional[bytes], Optional[bytes]]
+    ] = [  # type: ignore[assignment]
         (
             "http://scrapytest.org/1",  # parent
             "http://scrapytest.org/2",  # target